requirements.txt
fastapi
streamlit
uvicorn
requests
reportlab
python-pptx
transformers
mlflow
pytest
numpy
